@@ -26,8 +26,8 @@
         "symfony/security-bundle": "7.0.*",
         "symfony/translation": "7.0.*",
         "symfony/twig-bundle": "7.0.*",
-        "symfony/ux-live-component": "^2.13",
-        "symfony/ux-twig-component": "^2.13",
+        "symfony/ux-live-component": "^2.14",
+        "symfony/ux-twig-component": "^2.14",
         "symfony/validator": "7.0.*",
         "symfony/webpack-encore-bundle": "^2.1",
         "symfony/yaml": "7.0.*"
