@@ -6,6 +6,9 @@ use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
 use Symfony\UX\LiveComponent\DefaultActionTrait;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
 
+use App\Entity\Post;
+use Doctrine\ORM\EntityManagerInterface;
+
 #[AsLiveComponent]
 final class Search
 {
@@ -14,9 +17,16 @@ final class Search
     #[LiveProp(writable: true)]
     public string $query = '';
 
+    protected $entityManager;
+
+    public function __construct(EntityManagerInterface $entityManager)
+    {
+        $this->entityManager = $entityManager;
+    }
+
     public function getPosts(): array
     {
-        return $this->query ? [['title' => $this->query]] : [];
+        return $this->query ? $this->entityManager->getRepository(Post::class)->searchPosts($this->query) : [];
     }
 }
 
